requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
click==8.2.1
distro==1.9.0
fastapi==0.115.12
fpdf==1.7.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
lxml==5.4.0
numpy==2.2.6
openai==1.81.0
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
pydantic==2.11.4
pydantic_core==2.33.2
pytesseract==0.3.13
python-docx==1.1.2
python-multipart==0.0.20
qrcode==8.2
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
uvicorn==0.34.2
